refactored :peermetadata_interface
